Refactoring scrolling window approach for journald reading
The model only represents a visible data window of the journald
database, which is framed by a head and a tail cursor. The
cursors can be used to setup the journald reader to fetch
further data in a specific direction (head or tail).

This change makes jumping to the end of a journal possible and
will also allow to finally support journals that gain additional
data while the journal is open.