Refactor multi-threaded fetch access from QML
This commit fixes a very interesting integration with QML's ListView:
While sliding the visible area, muliple fetch() calls can be
triggered at the QAbstractItemModel. This can lead to situations
where the previous fetch did not complete yet and the beginInsertRows
and endInsertRows operations are wrongly ordered, because the QML
access are direct (and not queued).
This change introduces an atomic semaphore that ensures that at most one
fetch operation can be operative at once. For combination with model
reset this can be improved though by also using the same semaphor for
resets.