Correctly handle appending to log
In case of a non static log, e.g. one that is read from the current
system and which triggers updates, the FD listener shall trigger the
reading of more entries. In such cases, though, we are already at the
end of the log.
The only reasong for early skipping of further reading is a simple
performance gain. Thus, remove this check and update correctly.